threshold should be in milisec

+++ SDK-sources/OutbrainSDK/Models/Responses/OBSettings.h
@@ -28,7 +28,7 @@
 @property (nonatomic, copy, readonly, nullable) NSString *sourceFormat;
 
 @property (nonatomic, assign, readonly) BOOL isViewabilityPerListingEnabled;
-@property (nonatomic, assign, readonly) CGFloat viewabilityPerListingThresholdMilisec;
+@property (nonatomic, assign, readonly) NSInteger viewabilityPerListingThresholdMilisec;
 
 - (instancetype)initWithPayload:(NSDictionary *)payload;
